Moved score calculator to utils inside routes controllers
